fix: update error toast position and refactor components for improved structure, etc
