+++ src/pages/auth/login.tsx
@@ -32,7 +32,7 @@ const Login = () => {
       bgColor={"black"}
     >
       <Stack
-        minW={"320px"}
+        // minW={"320px"}
         maxW={"425px"}
         minH={"100vh"}
         bgColor={"#FAF9F6"}
@@ -91,9 +91,10 @@ const Login = () => {
             variant={"outline"}
             borderRadius={"lg"}
             borderColor={"#ECECEC"}
+            py={"1.5rem"}
             as={Link}
-            to={`https://api.clip-hub.tech/auth/login`}
-            // to={`http://localhost:8080/auth/login`}
+            // to={`https://api.clip-hub.tech/auth/login`}
+            to={`http://localhost:8080/auth/login`}
           >
             <Stack direction={"row"} align={"center"} justify={"center"}>
               <FcGoogle fontSize={"1.5rem"} />
@@ -123,6 +124,7 @@ const Login = () => {
               cursor={"pointer"}
               _hover={{ bgColor: "blackAlpha.800" }}
               onClick={connectToDevice}
+              py={"1.5rem"}
             >
               <Text fontSize={"0.8rem"} fontWeight={600} color={"#F0E13D"}>
                 Connect
